Restructured BaseMetaClass
